PMM-9849 fixed tests
