+++ public/app/percona/check/components/AllChecksTab/AllChecksTab.test.tsx
@@ -4,7 +4,7 @@ import { CheckService } from 'app/percona/check/Check.service';
 import { Interval } from 'app/percona/check/types';
 import { AllChecksTab } from './AllChecksTab';
 import { Messages } from './AllChecksTab.messages';
-import { render, screen, waitFor } from '@testing-library/react';
+import { fireEvent, render, screen, waitFor } from '@testing-library/react';
 
 jest.mock('@percona/platform-core', () => {
   const originalModule = jest.requireActual('@percona/platform-core');
@@ -17,17 +17,16 @@ jest.mock('@percona/platform-core', () => {
 });
 
 describe('AllChecksTab::', () => {
+  beforeEach(() => jest.clearAllMocks());
   it('should fetch checks at startup', async () => {
     const spy = jest.spyOn(CheckService, 'getAllChecks');
     render(<AllChecksTab />);
 
     expect(spy).toBeCalledTimes(1);
-
-    spy.mockClear();
   });
 
   it('should render a spinner at startup, while loading', async () => {
-    const spy = jest.spyOn(CheckService, 'getAllChecks').mockImplementation(() =>
+    jest.spyOn(CheckService, 'getAllChecks').mockImplementation(() =>
       Promise.resolve([
         {
           summary: 'Test',
@@ -42,11 +41,10 @@ describe('AllChecksTab::', () => {
     expect(screen.queryByTestId('spinner-wrapper')).toBeInTheDocument();
     await waitFor(() => component);
     expect(screen.queryByTestId('spinner-wrapper')).not.toBeInTheDocument();
-    spy.mockClear();
   });
 
   it('should log an error if the API call fails', async () => {
-    const spy = jest.spyOn(CheckService, 'getAllChecks').mockImplementation(() => {
+    jest.spyOn(CheckService, 'getAllChecks').mockImplementation(() => {
       throw Error('test');
     });
     const loggerSpy = jest.spyOn(logger, 'error').mockImplementationOnce(() => null);
@@ -54,8 +52,6 @@ describe('AllChecksTab::', () => {
     await waitFor(() => render(<AllChecksTab />));
 
     expect(loggerSpy).toBeCalledTimes(1);
-
-    spy.mockClear();
   });
 
   it('should render a table', async () => {
@@ -96,5 +92,48 @@ describe('AllChecksTab::', () => {
     expect(tbody.querySelectorAll('tr > td')[7]).toHaveTextContent(Messages.disabled);
     expect(tbody.querySelectorAll('tr > td')[8]).toHaveTextContent(Interval.RARE);
     expect(tbody.querySelectorAll('tr > td')[9]).toHaveTextContent(Messages.enable);
+  });
+
+  it('should log an error if the run checks API call fails', async () => {
+    jest.spyOn(CheckService, 'runDbChecks').mockImplementationOnce(() => {
+      throw Error('test');
+    });
+    const loggerSpy = jest.spyOn(logger, 'error');
+
+    render(<AllChecksTab />);
+
+    await screen.findByTestId('db-check-panel-actions');
+
+    const runChecksButton = screen.getByRole('button', { name: Messages.runDbChecks });
+
+    await waitFor(() => fireEvent.click(runChecksButton));
+    fireEvent.click(runChecksButton);
+    expect(screen.queryByText('Run Checks')).not.toBeInTheDocument();
+
+    await waitFor(() => {
+      expect(loggerSpy).toBeCalledTimes(1);
+    });
+
+    expect(await screen.findByText('Run Checks')).toBeInTheDocument();
+  });
+
+  it('should call the API to run checks when the "run checks" button gets clicked', async () => {
+    const runChecksSpy = jest.spyOn(CheckService, 'runDbChecks');
+    await render(<AllChecksTab />);
+
+    await screen.findByTestId('db-check-panel-actions');
+
+    const runChecksButton = screen.getByRole('button', { name: Messages.runDbChecks });
+
+    expect(runChecksSpy).toBeCalledTimes(0);
+    fireEvent.click(runChecksButton);
+
+    expect(screen.queryByText('Run Checks')).not.toBeInTheDocument();
+
+    await waitFor(() => {
+      expect(runChecksSpy).toBeCalledTimes(1);
+    });
+
+    expect(await screen.findByText('Run Checks')).toBeInTheDocument();
   });
 });
